Reorganize tests, remove doc comments
