+++ cobalt-parser/src/lexer/mod.rs
@@ -77,18 +77,3 @@ impl<'src> SourceReader<'src> {
         SourceSpan::from((offset, self.index - offset))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_slice_backward() {
-        let source = "hello world";
-        let mut reader = SourceReader::new(source);
-        reader.next_char();
-        reader.next_char();
-        reader.next_char();
-        assert_eq!(reader.slice_backward(3), "hel");
-    }
-}
