+++ cobalt-parser/src/parser/expr/binop_rhs.rs
@@ -29,7 +29,7 @@ impl<'src> Parser<'src> {
     /// - After eating the '(', it calls the parse expr so as before we enter this
     /// function, now with the lhs being 'b' with precedence 0 and again pointing to
     /// a `+` token. We eat it, parse the `c` as a primary expression.
-    pub fn parse_binop_rhs(
+    pub(crate) fn parse_binop_rhs(
         &mut self,
         lhs_precedence: u8,
         mut lhs: BoxedAST<'src>,
