+++ cobalt-parser/src/tests/inputs/test_2.co
@@ -1,0 +1,20 @@
+module alloc._utils;
+@link(weak) @forward
+fn memclear(ptr: *mut null, size: usize) = {
+  let ptr = ptr: *mut u8;
+  let mut i = 0;
+  while (i < size) {
+    ptr[i] = 0;
+    ++i;
+  }
+};
+@link(weak) @forward
+fn memcpy(dst: *mut null, src: *null, size: usize) = {
+  let dst = dst: *mut u8;
+  let src = src: *u8;
+  let mut i = 0;
+  while (i < size) {
+    dst[i] = src[i];
+    ++i;
+  }
+};
