+++ cobalt-parser/src/tests/mod.rs
@@ -1,0 +1,4 @@
+use super::*;
+
+mod lexer;
+mod parser;
